# apiGateway/api/docs/swagger.yaml
definitions:
  models.Account:
    properties:
      balance:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.AccountsResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/models.Account'
        type: array
      count:
        type: integer
    type: object
  models.Budget:
    properties:
      amount:
        type: number
      categoryId:
        type: string
      createdAt:
        type: string
      endTime:
        type: string
      id:
        type: string
      period:
        type: string
      startTime:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.BudgetsResponse:
    properties:
      budgets:
        items:
          $ref: '#/definitions/models.Budget'
        type: array
      count:
        type: integer
    type: object
...
  models.CreateAccount:
    properties:
      balance:
        type: number
      currency:
        type: string
      name:
        type: string
      type:
        type: string
      userId:
        type: string
    type: object
  models.CreateBudget:
    properties:
      amount:
        type: number
      categoryId:
        type: string
      endTime:
        type: string
      period:
        type: string
      startTime:
        type: string
      userId:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateAccount:
    properties:
      balance:
        type: number
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      userId:
        type: string
    type: object
  models.UpdateBudget:
    properties:
      amount:
        type: number
      categoryId:
        type: string
      endTime:
        type: string
      id:
        type: string
      period:
        type: string
      startTime:
        type: string
      userId:
        type: string
    type: object
...
info:
  contact: {}
  description: Personal Finance Tracker
  title: Personal Finance Tracker API
  version: "1"
paths:
  /budget_service/v1/account:
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new account
      tags:
      - account
  /budget_service/v1/account/{id}:
    delete:
      consumes:
      - application/json
      description: Delete account
      parameters:
      - description: account_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete account
      tags:
      - account
    get:
      consumes:
      - application/json
      description: Get account by id
      parameters:
      - description: account_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get account by id
      tags:
      - account
    put:
      consumes:
      - application/json
      description: Update account
      parameters:
      - description: account_id
        in: path
        name: id
        required: true
        type: string
      - description: account
        in: body
        name: account
        schema:
          $ref: '#/definitions/models.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update account
      tags:
      - account
  /budget_service/v1/accounts:
    get:
      consumes:
      - application/json
      description: Get all account
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all account
      tags:
      - account
  /budget_service/v2/budget:
    post:
      consumes:
      - application/json
      description: Create a new budget
      parameters:
      - description: budget
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/models.CreateBudget'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Budget'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new budget
      tags:
      - budget
  /budget_service/v2/budget/{id}:
    delete:
      consumes:
      - application/json
      description: Delete budget
      parameters:
      - description: budget_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete budget
      tags:
      - budget
    get:
      consumes:
      - application/json
      description: Get budget by id
      parameters:
      - description: budget_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Budget'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get budget by id
      tags:
      - budget
    put:
      consumes:
      - application/json
      description: Update budget
      parameters:
      - description: budget_id
        in: path
        name: id
        required: true
        type: string
      - description: budget
        in: body
        name: budget
        schema:
          $ref: '#/definitions/models.UpdateBudget'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Budget'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update budget
      tags:
      - budget
  /budget_service/v2/budgets:
    get:
      consumes:
      - application/json
      description: Get all budgets
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BudgetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all budgets
      tags:
      - budget
...
swagger: "2.0"